requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.6
django-admin-interface==0.30.1
django-colorfield==0.14.0
django-unfold==0.65.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==11.3.0
python-slugify==8.0.4
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.15.0
uritemplate==4.2.0
